requirements.txt
python-dotenv
pycli2
requests
